+++ packages/router/src/lib/helpers/shared.ts
@@ -4,6 +4,7 @@ import {
   MethodContext,
   GAS_ESTIMATES,
   getChainData,
+  createLoggingContext,
 } from "@connext/nxtp-utils";
 import { BigNumber, constants, utils } from "ethers";
 
@@ -13,7 +14,7 @@ import { getContext } from "../../router";
 import { SwapValidInput } from "../entities";
 import { SwapInvalid } from "../errors";
 
-const NO_ORACLE_CHAINS = [10];
+const NO_ORACLE_CHAINS: number[] = [];
 
 /**
  * Helper to allow easy mocking
@@ -76,11 +77,19 @@ export const calculateGasFeeInReceivingToken = async (
   receivingAssetId: string,
   receivingChainId: number,
   outputDecimals: number,
-  requestContext: RequestContext,
+  _requestContext: RequestContext,
 ): Promise<BigNumber> => {
-  // NOTE: hardcoding in optimism to allow for fees before oracle can be
-  // properly deployed
-  const chaindIdsForGasFee = [...getChainIdForGasFee(), 10];
+  const { logger } = getContext();
+  const { requestContext, methodContext } = createLoggingContext(calculateGasFeeInReceivingToken.name, _requestContext);
+  logger.info("Method start", requestContext, methodContext, {
+    sendingChainId,
+    sendingAssetId,
+    receivingAssetId,
+    receivingChainId,
+    outputDecimals,
+  });
+
+  const chaindIdsForGasFee = getChainIdForGasFee();
 
   if (!chaindIdsForGasFee.includes(sendingChainId) && !chaindIdsForGasFee.includes(receivingChainId))
     return constants.Zero;
@@ -104,6 +113,13 @@ export const calculateGasFeeInReceivingToken = async (
     ? await getMainnetEquivalent(receivingAssetId, receivingChainId)
     : receivingAssetId;
 
+  logger.info("Getting token prices", requestContext, methodContext, {
+    tokenPricingSendingChain,
+    tokenPricingReceivingChain,
+    tokenPricingAssetIdSendingChain,
+    tokenPricingAssetIdReceivingChain,
+    outputDecimals,
+  });
   if (chaindIdsForGasFee.includes(sendingChainId)) {
     const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
     const [ethPriceInSendingChain, receivingTokenPrice, gasPriceInSendingChain] = await Promise.all([
@@ -123,6 +139,13 @@ export const calculateGasFeeInReceivingToken = async (
       ? constants.Zero
       : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
     totalCost = totalCost.add(tokenAmountForGasFee);
+    logger.info("Calculated cost on sending chain", requestContext, methodContext, {
+      totalCost: totalCost.toString(),
+      l1GasInUsd: l1GasInUsd.toString(),
+      ethPriceInSendingChain: ethPriceInSendingChain.toString(),
+      receivingTokenPrice: receivingTokenPrice.toString(),
+      gasPriceInSendingChain: gasPriceInSendingChain.toString(),
+    });
   }
 
   if (chaindIdsForGasFee.includes(receivingChainId)) {
@@ -149,10 +172,17 @@ export const calculateGasFeeInReceivingToken = async (
       : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
 
     totalCost = totalCost.add(tokenAmountForGasFee);
+    logger.info("Calculated cost on receiving chain", requestContext, methodContext, {
+      totalCost: totalCost.toString(),
+      tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+      l1GasInUsd: l1GasInUsd.toString(),
+      ethPriceInSendingChain: ethPriceInReceivingChain.toString(),
+      receivingTokenPrice: receivingTokenPrice.toString(),
+      gasPriceInSendingChain: gasPriceInReceivingChain.toString(),
+    });
   }
 
   // convert back to the intended decimals
-  // return totalCost.div(BigNumber.from(10).pow(18 - _outputDecimals));
   return totalCost;
 };
 
@@ -168,25 +198,57 @@ export const calculateGasFeeInReceivingTokenForFulfill = async (
   receivingAssetId: string,
   receivingChainId: number,
   outputDecimals: number,
-  requestContext: RequestContext,
+  _requestContext: RequestContext,
 ): Promise<BigNumber> => {
+  const { logger } = getContext();
+  const { requestContext, methodContext } = createLoggingContext(calculateGasFeeInReceivingToken.name, _requestContext);
+  logger.info("Method start", requestContext, methodContext, {
+    receivingAssetId,
+    receivingChainId,
+    outputDecimals,
+  });
   const chaindIdsForGasFee = getChainIdForGasFee();
 
   if (!chaindIdsForGasFee.includes(receivingChainId)) return constants.Zero;
   let totalCost = constants.Zero;
 
+  const tokenPricingReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId) ? 1 : receivingChainId;
+
+  const tokenPricingAssetIdReceivingChain = NO_ORACLE_CHAINS.includes(receivingChainId)
+    ? await getMainnetEquivalent(receivingAssetId, receivingChainId)
+    : receivingAssetId;
+
   if (chaindIdsForGasFee.includes(receivingChainId)) {
     const gasLimitForFulfill = BigNumber.from(GAS_ESTIMATES.fulfill);
     const [ethPriceInReceivingChain, receivingTokenPrice, gasPriceInReceivingChain] = await Promise.all([
-      getTokenPrice(receivingChainId, constants.AddressZero, requestContext),
-      getTokenPrice(receivingChainId, receivingAssetId, requestContext),
+      getTokenPrice(tokenPricingReceivingChain, constants.AddressZero, requestContext),
+      getTokenPrice(tokenPricingReceivingChain, tokenPricingAssetIdReceivingChain, requestContext),
       getGasPrice(receivingChainId, requestContext),
     ]);
 
-    const gasAmountInUsd = gasPriceInReceivingChain.mul(gasLimitForFulfill).mul(ethPriceInReceivingChain);
+    // https://community.optimism.io/docs/users/fees-2.0.html#fees-in-a-nutshell
+    let l1GasInUsd = BigNumber.from(0);
+    if (receivingChainId === 10) {
+      const gasPriceMainnet = await getGasPrice(1, requestContext);
+      l1GasInUsd = gasPriceMainnet.mul(GAS_ESTIMATES.prepareL1).mul(ethPriceInReceivingChain);
+    }
+
+    const gasAmountInUsd = gasPriceInReceivingChain
+      .mul(gasLimitForFulfill)
+      .mul(ethPriceInReceivingChain)
+      .add(l1GasInUsd);
     const tokenAmountForGasFee = receivingTokenPrice.isZero()
       ? constants.Zero
       : gasAmountInUsd.div(receivingTokenPrice).div(BigNumber.from(10).pow(18 - outputDecimals));
+
+    logger.info("Calculated cost on receiving chain for fulfill", requestContext, methodContext, {
+      totalCost: totalCost.toString(),
+      tokenAmountForGasFee: tokenAmountForGasFee.toString(),
+      l1GasInUsd: l1GasInUsd.toString(),
+      ethPriceInReceivingChain: ethPriceInReceivingChain.toString(),
+      receivingTokenPrice: receivingTokenPrice.toString(),
+      gasPriceInReceivingChain: gasPriceInReceivingChain.toString(),
+    });
 
     totalCost = totalCost.add(tokenAmountForGasFee);
   }