+++ packages/utils/src/asset.ts
@@ -1,6 +1,7 @@
 import { providers, BigNumber, Contract, constants } from "ethers";
 
 import { ERC20Abi } from "./basic";
+import { ChainData } from "./chainData";
 
 export const getOnchainBalance = async (
   assetId: string,
@@ -10,4 +11,21 @@ export const getOnchainBalance = async (
   return assetId === constants.AddressZero
     ? provider.getBalance(address)
     : new Contract(assetId, ERC20Abi, provider).balanceOf(address);
+};
+
+export const getDecimalsForAsset = async (
+  assetId: string,
+  chainId: number,
+  provider: providers.Provider,
+  chainData?: Map<string, ChainData>,
+): Promise<number> => {
+  if (chainData) {
+    const chainInfo = chainData.get(chainId.toString());
+    const decimals = chainInfo?.assetId[assetId]?.decimals;
+    if (decimals) {
+      return decimals;
+    }
+  }
+  const contract = new Contract(assetId, ERC20Abi, provider);
+  return await contract.decimals();
 };